requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.9.1
babel==2.17.0
boto3==1.39.15
botocore==1.39.15
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cssselect2==0.8.0
defusedxml==0.7.1
Django==4.2.23
django-countries==7.6.1
django-formtools==2.5.1
django-picklefield==3.3
django-q2==1.7.4
django-storages==1.14.6
django-tenants==3.8.0
et_xmlfile==2.0.0
gunicorn==23.0.0
html5lib==1.1
idna==3.10
imgkit==1.2.3
importlib_metadata==8.7.0
jmespath==1.0.1
minio==7.2.16
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
psycopg2==2.9.9
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.23.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.4
s3transfer==0.13.1
sib-api-v3-sdk==7.6.0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.20
WeasyPrint==52.5
webencodings==0.5.1
zipp==3.23.0
